Add gradle dependency management
