+++ .gitignore
@@ -1,3 +1,6 @@
 out/
 .idea/
 *.iml
+
+build/
+.gradle/
